app/requirements.txt
python-telegram-bot==21.10
ollama==0.4.5
requests